refactor: update button imports and improve component structure; adjust props for consistency
