+++ package.json
@@ -9,7 +9,7 @@
         "coverage": "vitest run --coverage",
         "dev": "next dev",
         "dev:turbo": "next dev --turbo",
-        "lint": "biome lint --apply .",
+        "lint": "biome lint --write .",
         "start": "next start",
         "storybook": "storybook dev -p 6006",
         "test": "vitest",
@@ -31,7 +31,6 @@
         "@radix-ui/react-dropdown-menu": "^2.1.16",
         "@radix-ui/react-icons": "1.3.2",
         "@radix-ui/react-slot": "^1.2.3",
-        "@tabler/icons-react": "^3.35.0",
         "@tailwindcss/forms": "0.5.10",
         "@tanstack/react-query": "5.87.4",
         "@tanstack/react-query-devtools": "5.87.4",
