+++ stories/Button.stories.tsx
@@ -1,6 +1,7 @@
 import { StoryFn } from '@storybook/react';
+import { Button, ButtonProps } from '../components/ui/button';
 
-import Button, { ButtonProps } from '../components/Button';
+
 
 export default {
     title: 'Button',
