+++ tests/Button.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import { describe, test } from 'vitest';
+import { Button } from '../components/ui/button';
 
-import Button from '../components/Button';
 
 /**
  * @jest-environment jsdom
